requirements.txt
fastapi
uvicorn[standard]
PyGithub
openai
requests
python-dotenv
pydantic